refactor: move Content and Part from model package to dto package
