+++ src/main/java/com/example/InsightEngine/dto/Part.java
@@ -1,4 +1,4 @@
-package com.example.InsightEngine.model;
+package com.example.InsightEngine.dto;
 
 public class Part {
     private String text;
